+++ lib/Doctrine/ODM/PHPCR/Document/AbstractFile.php
@@ -106,16 +106,6 @@ abstract class AbstractFile implements HierarchyInterface
     }
 
     /**
-     * Kept for BC
-     *
-     * @deprecated use getParentDocument instead.
-     */
-    public function getParent()
-    {
-        return $this->getParentDocument();
-    }
-
-    /**
      * Set the parent document of this document.
      *
      * @param object $parent Document that is the parent of this node. Could be
@@ -128,16 +118,6 @@ abstract class AbstractFile implements HierarchyInterface
         $this->parent = $parent;
 
         return $this;
-    }
-
-    /**
-     * Kept for BC
-     *
-     * @deprecated use setParentDocument instead.
-     */
-    public function setParent($parent)
-    {
-        return $this->setParentDocument($parent);
     }
 
     /**