+++ lib/Doctrine/ODM/PHPCR/Document/Resource.php
@@ -117,16 +117,6 @@ class Resource
     }
 
     /**
-     * Kept for BC
-     *
-     * @deprecated use getParentDocument instead.
-     */
-    public function getParent()
-    {
-        return $this->getParentDocument();
-    }
-
-    /**
      * Set the parent document of this resource.
      *
      * @param object $parent Document that is the parent of this node.
@@ -138,17 +128,6 @@ class Resource
         $this->parent = $parent;
 
         return $this;
-    }
-
-
-    /**
-     * Kept for BC
-     *
-     * @deprecated use setParentDocument instead.
-     */
-    public function setParent($parent)
-    {
-        return $this->setParentDocument($parent);
     }
 
     /**