+++ lib/Doctrine/ODM/PHPCR/Mapping/ClassMetadata.php
@@ -1322,15 +1322,6 @@ class ClassMetadata implements ClassMetadataInterface
 
     /**
      * {@inheritDoc}
-     * @deprecated use getFieldMapping instead
-     */
-    public function getField($fieldName)
-    {
-        return $this->getFieldMapping($fieldName);
-    }
-
-    /**
-     * {@inheritDoc}
      */
     public function hasAssociation($fieldName)
     {