+++ CHANGELOG.md
@@ -4,13 +4,15 @@ Changelog
 2.0.0 (not released)
 --------------------
 
-* **2017-01-15** [Feature] `DocumentManager::getDocumentId()` to get the id of a managed document
+* **2017-01-17**: [BC Break] removed deprecated methods HierarchyInterface:.getParent/setParent. 
+                  DocumentManager::flush no longer saves the PHPCR session if there are no changes on the ODM layer.
+                  Removed deprecated annotations for fields. Use `@Field(type="...")` instead.
+* **2017-01-15**: [Feature] `DocumentManager::getDocumentId()` to get the id of a managed document
 
 1.4.0
 -----
 
-* **2016-06-08**  [Feature] Allow children of a document to be restricted to a
-                  certain class or forbidden.
+* **2016-06-08**: [Feature] Allow children of a document to be restricted to a certain class or forbidden.
 
 1.3.1
 -----